Preserve aspect ratio in fullscreen mode.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource/Main@1365 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 116fe293b3a9d63239991549cd64fe7c2145c302

+++ Source/SoundManager.cpp
@@ -548,7 +548,7 @@ bool SoundManager::ConvertData( Sound* sound )
     if( r == 1 )
     {
         cvt.len = sound->BaseBufSize;
-        if( cvt.len * cvt.len_mult > sound->ConvertedBufRealSize )
+        if( cvt.len * cvt.len_mult > (int) sound->ConvertedBufRealSize )
         {
             sound->ConvertedBufRealSize = cvt.len * cvt.len_mult;
             SAFEDELA( sound->ConvertedBuf );
